Refactor Element into partials
